good

+++ Libft/ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: mpinthas <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/01 14:36:05 by mpinthas          #+#    #+#             */
-/*   Updated: 2024/09/01 14:38:41 by mpinthas         ###   ########.fr       */
+/*   Updated: 2024/09/01 17:41:31 by mpinthas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
@@ -15,8 +15,8 @@ int	ft_memcmp(const void *str1, const void *str2, size_t n)
 {
 	char	*s1;
 	char	*s2;
-	int	i;
-	
+	size_t	i;
+
 	s1 = (char *)str1;
 	s2 = (char *)str2;
 	i = 0;
